% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_score.R
\name{add_scores}
\alias{add_scores}
\title{This functions adds the score of a given item set and a key set using the psych package}
\usage{
add_scores(df, item_prefix, keys)
}
\arguments{
\item{df}{the dataframe to use}

\item{item_prefix}{the prefix common to all items. Data is selected using dplyr::starts_with}

\item{keys}{a keylist as used in the psych::scoreItems function}
}
\value{
the data frame with an added column
}
\description{
This functions adds the score of a given item set and a key set using the psych package
}
