% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{add_survey_response_duration}
\alias{add_survey_response_duration}
\title{Add a survey response duration to surveymonkey data-frames.
Dataframes must contain a date_created and date_modified column.}
\usage{
add_survey_response_duration(df)
}
\arguments{
\item{df}{Surveymonkey dataframe}
}
\value{
A dataframe with new column for survey duration
}
\description{
Add a survey response duration to surveymonkey data-frames.
Dataframes must contain a date_created and date_modified column.
}
