% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{as.numeric.factor}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{Vector that contains a factor.}
}
\value{
Vector that contais numerics.
}
\description{
Returns a numerical vector that was a factor vector. Is used to create means from likert data.
}
