% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_score.R
\name{auto_score}
\alias{auto_score}
\title{This function automatically scores a set of items and creates a scale variable.
It detects inverse coding using the psych::alpha function.}
\usage{
auto_score(df, prefix)
}
\arguments{
\item{df}{a dataframe that contains an item set}

\item{prefix}{the prefix used in all items and also the name of the new scale.}
}
\value{
the dataframe with the new scale column
}
\description{
This function automatically scores a set of items and creates a scale variable.
It detects inverse coding using the psych::alpha function.
}
