% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataUtils.R
\name{copy_labels}
\alias{copy_labels}
\title{Copy labels from one data set to another}
\usage{
copy_labels(from, to)
}
\arguments{
\item{from}{the source of the labels}

\item{to}{the target of the labels}
}
\value{
the labeled tibble
}
\description{
This is often used when merging data sets to ensure that the labels, still exist
after bind_rows
}
