% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciPlots.R
\name{createData}
\alias{createData}
\title{helper function that creates the data for the printCI call}
\usage{
createData(lowerMean, upperMean, lowerSE, upperSE, diffLo, diffHigh, labelA,
  labelB, title)
}
\description{
helper function that creates the data for the printCI call
}
