% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw.palette}
\alias{draw.palette}
\title{draw.palette}
\usage{
draw.palette(col, border = "light gray", ...)
}
\arguments{
\item{col}{Palette vector}

\item{border}{Border used for plotting}

\item{...}{Additional params forwarded to plot.}
}
\description{
Utility function that draws a palette.
}
