% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Generate a Code-Book from a surveymonkey data-frame}
\usage{
generate_codebook(df, filename)
}
\arguments{
\item{df}{Surveymonkey dataframe}

\item{filename}{Filename for the codebook (xls compatible format)}
}
\value{
nothing
}
\description{
Generate a Code-Book from a surveymonkey data-frame
}
