% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalPlots.R
\name{get.delta.range}
\alias{get.delta.range}
\title{get.delta.range}
\usage{
get.delta.range(data, steps = 4)
}
\arguments{
\item{data}{Data that contains 3 lower and upper variables}

\item{steps}{Size of the stride}
}
\value{
a range
}
\description{
Utility function to generate the delta axis
}
