% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovi_plots.R
\name{get_emm_plot}
\alias{get_emm_plot}
\title{Get ggplot object from jmv::ANOVA}
\usage{
get_emm_plot(jmv_result, plot_number = 1)
}
\arguments{
\item{jmv_result}{the result of a jmv::ANOVA call}

\item{plot_number}{the index of the emm plot to retrieve}
}
\value{
Returns a ggplot object
}
\description{
This function provides access to the emm plots of a jamovi ANOVA object
}
\examples{
jmv_result <- jmv::ANOVA(
   data = ToothGrowth,
   dep = "len",
   factors = c("supp", "dose"),
   emMeans = list(
      c("dose", "supp"))
      )

get_emm_plot(jmv_result)
}
