% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciPlots.R
\name{getDeltaRange}
\alias{getDeltaRange}
\title{return the delta-ranges for automatic delta axis size}
\usage{
getDeltaRange(data, steps = 4)
}
\description{
return the delta-ranges for automatic delta axis size
}
