% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciPlots.R
\name{independentSampleMeansCI}
\alias{independentSampleMeansCI}
\title{print a CI plot from two different means}
\usage{
independentSampleMeansCI(data1, data2, labelA = "A", labelB = "B",
  title = "Dependent Variable", range = c(1, 0))
}
\description{
print a CI plot from two different means
}
