% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{load_surveymonkey_csv}
\alias{load_surveymonkey_csv}
\title{Loads surveymonkey generated csv files that contain 2 line headers}
\usage{
load_surveymonkey_csv(filename)
}
\arguments{
\item{filename}{Character of the file location.}
}
\value{
raw data with a single header
}
\description{
Loads surveymonkey generated csv files that contain 2 line headers
}
