% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalPlots.R
\name{mean_se}
\alias{mean_se}
\title{mean.se}
\usage{
mean_se(x)
}
\arguments{
\item{x}{data}
}
\value{
Standard error
}
\description{
Returns the standard error of a vector
}
\examples{
mean.se(c(1, 2, 3, 4, 5, 6))
}
