% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeType}
\alias{mergeType}
\title{merges two datasets and determine ordering or not by parameter
must ensure both datasets are compatible}
\usage{
mergeType(d1, d2, ordered = FALSE)
}
\description{
merges two datasets and determine ordering or not by parameter
must ensure both datasets are compatible
}
