% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{name_variables}
\alias{name_variables}
\title{Rename all variables by a character vector, but keep unnamed variables the same}
\usage{
name_variables(df, columnvector)
}
\arguments{
\item{df}{The dataframe to rename}

\item{columnvector}{A vector of names}
}
\value{
A dataframe with renames columns
}
\description{
Rename all variables by a character vector, but keep unnamed variables the same
}
