% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careless_index.R
\name{numbers_only}
\alias{numbers_only}
\title{Tests if a vector of strings can be a numeric}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{a vector of strings}
}
\value{
Returns a vector of booleans
}
\description{
Tests if a vector of strings can be a numeric
}
