% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalPlots.R
\name{plot.CI}
\alias{plot.CI}
\title{plot.CI Plots two independent sample CI}
\usage{
\method{plot}{CI}(d, titleScale = "Dependent Variable", delta, range = c(1,
  6), autorange = TRUE)
}
\arguments{
\item{d}{Data Frame}

\item{titleScale}{Name of the Scale for the dependent Varialbe}

\item{delta}{a}

\item{range}{Axis Ranges}

\item{autorange}{}
}
\value{

}
\description{
plot.CI Plots two independent sample CI
}
