% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalPlots.R
\name{plot_IS_meansCI}
\alias{plot_IS_meansCI}
\title{plot.IS.meansCI}
\usage{
plot_IS_meansCI(
  x,
  y,
  labelA = "A",
  labelB = "B",
  title = "Dependent Variable",
  range = c(1, 0)
)
}
\arguments{
\item{x}{vector of sample a}

\item{y}{vector of sample b}

\item{labelA}{label for variable a}

\item{labelB}{label for variable b}

\item{title}{Title for the plot}

\item{range}{Adjust the range of the plot}
}
\value{
plots
}
\description{
plot.IS.meansCI
}
\examples{
plot_IS_meansCI(c(1, 2, 3), c(3, 4, 5))

}
