% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciPlots.R
\name{printCI}
\alias{printCI}
\title{Print a CI Interval for a given set of data}
\usage{
printCI(d, titleScale = "Dependent Variable", delta, range = c(1, 6),
  autorange = TRUE)
}
\description{
Print a CI Interval for a given set of data
}
