% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{read_codebook}
\alias{read_codebook}
\title{Read a codebook file for a surveymonkey dataframe}
\usage{
read_codebook(filename)
}
\arguments{
\item{filename}{Filename for the codebook (xls compatible format)}
}
\value{
nothing
}
\description{
Read a codebook file for a surveymonkey dataframe
}
