% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgb2hex}
\alias{rgb2hex}
\title{converts rgb values to hex notation}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r}{red value}

\item{g}{green value}

\item{b}{blue value}
}
\description{
converts rgb values to hex notation
}
