% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robot_care}
\alias{robot_care}
\title{Attitudes towards robotic care.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{age}{age of participant, in years}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
Self-Collected \url{http://www.diamondse.info/}
}
\usage{
robot_care
}
\description{
A dataset containing the attitudes of participants to robotic care.
}
\keyword{datasets}
