% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/careless_index.R
\name{careless_indices}
\alias{careless_indices}
\title{Detect careless responses}
\usage{
careless_indices(
  dat,
  speeder_analysis = F,
  likert_vector = NULL,
  duration_column = "Duration (in seconds)",
  id_column = "ResponseId",
  antonyms = TRUE,
  critval_ant = -0.6,
  synonyms = TRUE,
  critval_syn = 0.6,
  mahaD = TRUE
)
}
\arguments{
\item{dat}{A data.frame that contains the survey data.}

\item{speeder_analysis}{A criterion to test against speeders. The
speeder_analysis parameter can be one of the following: median,
median/2 (default), median/3, or a numeric value that is directly used
as a threshold. Using FALSE will ignore speeder analysis.}

\item{likert_vector}{A vector of indices that should be analyzed.}

\item{duration_column}{A string that contains the name of the column that
has the duration of survey responses. As a default we use
Duration (in seconds), which is used by Qualtrix.}

\item{id_column}{A string that contains the name of a id column. As a default
we use "ResponseId", which is used by Qualtrix.}

\item{antonyms}{Whether or not to test for psychometric antonyms (default: T)}

\item{critval_ant}{Threshold for antonym analysis (default: -0.6)}

\item{synonyms}{Whether or not to test for psychometric synonyms (default: T)}

\item{critval_syn}{Threshold for synonym analysis (default: -0.6)}

\item{mahaD}{Whether or not to test for mahalanobis distance with confidence of 0.99. (default: T)}
}
\value{
Returns an updated dataframe containing indices
}
\description{
A function to detect careless survey responses using the careless package.
The function has the improvement to first remove speeders, because they
negatively affect the metrics used in the careless package.
}
