% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlots.R
\name{cor.matrix.plot}
\alias{cor.matrix.plot}
\title{Plots a correlation matrix plot.}
\usage{
cor.matrix.plot(data, conf.level = 0.95)
}
\arguments{
\item{data}{Data frame to plot. Use `dplyr::select` to filter variables to use.}

\item{conf.level}{Level of significance (default .95)}
}
\value{
The function is called for its side-effects (plot).
}
\description{
This function creates a plot of correlations with sensible defaults.
}
\examples{
cars \%>\% dplyr::select(speed, dist) \%>\% cor.matrix.plot()

}
