% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlots.R
\name{cor.matrix.plot}
\alias{cor.matrix.plot}
\title{cor.matrix.plot}
\usage{
cor.matrix.plot(data)
}
\arguments{
\item{data}{Data frame to plot. Use dplyer::select to filter variables to use.}

\item{conf.level}{Level of significance (default .95)}
}
\value{

}
\description{
Plots a correlation matrix plot. The to
}
\examples{
cars \%>\% dplyr::select(speed, dist) \%>\% cor.matrix.plot()

}
