% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_type}
\alias{merge_type}
\title{merge.type}
\usage{
merge_type(d1, d2, ordered = FALSE)
}
\arguments{
\item{d1}{first data frame}

\item{d2}{second data frame}

\item{ordered}{}
}
\value{

}
\description{
Merges two dataframes
}
