% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalPlots.R
\name{prepare.data}
\alias{prepare.data}
\title{prepare.data}
\usage{
prepare.data(lowerMean, upperMean, lowerSE, upperSE, diffLo, diffHigh, labelA,
  labelB, title)
}
\arguments{
\item{title}{}
}
\value{

}
\description{
Utility function used to create a data.frame taht contains the values needed for a CI plot.
}
